Tidied up authorization around follows
