+++ app/controllers/follows_controller.rb
@@ -1,8 +1,10 @@
 class FollowsController < ApplicationController
+  before_filter :authenticate_member!
+  load_and_authorize_resource
 
   # POST /follows
   def create
-    
+
     @follow = current_member.follows.build(:followed_id => follow_params[:followed_id])
 
     if @follow.save
