+++ app/views/members/show.html.haml
@@ -5,14 +5,16 @@
     = link_to 'Edit profile', edit_member_registration_path, :class => 'btn btn-default'
   - if @member == current_member && !@member.is_paid?
     = link_to "Upgrade account", shop_path, :class => 'btn btn-default'
-  -if can? :create, Notification and current_member != @member
-    =link_to 'Send message', new_notification_path(:recipient_id => @member.id), :class => 'btn btn-default'
-  - if can? :create, Follow and current_member != @member
-    - if !current_member.already_following?(@member)
+  - if can? :create, Notification and current_member != @member
+    = link_to 'Send message', new_notification_path(:recipient_id => @member.id), :class => 'btn btn-default'
+
+  - if current_member and current_member != @member # must be logged in, can't follow yourself
+    - follow = current_member.get_follow(@member)
+    - if !follow and can? :create, Follow # not already following
       = link_to 'Follow', follows_path(:followed_id => @member.id), :method => :post, :class => 'btn btn-default'
-    - else
-      - follow = current_member.get_follow(@member)
+    - if follow and can? :destroy, follow # already following
       = link_to 'Unfollow', follow_path(follow), :method => :delete, :class => 'btn btn-default'
+
 - content_for :member_rss_login_name, @member.login_name
 - content_for :member_rss_slug, @member.slug
