+++ spec/features/member_profile_spec.rb
@@ -13,8 +13,6 @@ feature "member profile" do
       expect(page).to have_content "Account type: Free account"
       expect(page).to have_content "#{member.login_name}'s gardens"
       expect(page).to have_link "More about this garden...", :href => garden_path(member.gardens.first)
-      expect(page).to_not have_link "Follow"
-      expect(page).to_not have_link "Unfollow"
     end
 
     scenario "no bio" do
@@ -144,52 +142,6 @@ feature "member profile" do
 
       scenario "has a private message button" do
         expect(page).to have_link "Send message", :href => new_notification_path(:recipient_id => other_member.id)
-      end
-
-      scenario "has a follow button" do
-        expect(page).to have_link "Follow", :href => follows_path(:followed_id => other_member.id)
-      end
-    end
-
-    context "following another member" do
-      background do
-        visit member_path(other_member)
-        click_link 'Follow'
-      end
-
-      scenario "has correct message and unfollow button" do
-        expect(page).to have_content "Followed #{other_member.login_name}"
-        expect(page).to have_link "Unfollow", :href => follow_path(member.get_follow(other_member))
-      end
-
-      scenario "has a followed member listed in the following page" do
-        visit member_follows_path(member)
-        expect(page).to have_content "#{other_member.login_name}"
-      end
-
-      scenario "has correct message and follow button after unfollow" do
-        click_link 'Unfollow'
-        expect(page).to have_content "Unfollowed #{other_member.login_name}"
-        visit member_path(other_member) # unfollowing redirects to root
-        expect(page).to have_link "Follow", :href => follows_path(:followed_id => other_member.id)
-      end
-
-      scenario "has member in following list" do
-        visit member_follows_path(member)
-        expect(page).to have_content "#{other_member.login_name}"
-      end
-
-      scenario "appears in in followed member's followers list" do
-        visit member_followers_path(other_member)
-        expect(page).to have_content "#{member.login_name}"
-      end
-
-      scenario "removes members from following and followers lists after unfollow" do
-        click_link 'Unfollow'
-        visit member_follows_path(member)
-        expect(page).not_to have_content "#{other_member.login_name}"
-        visit member_followers_path(other_member)
-        expect(page).not_to have_content "#{member.login_name}"
       end
 
     end
